Refactor: 재사용 아이콘 통합
